moving fb pixel statics
